% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.welfare.R
\name{print.welfare}
\alias{print.welfare}
\title{Print a formatted welfare decomposition}
\usage{
\method{print}{welfare}(x, label = "Welfare Change", ...)
}
\arguments{
\item{x}{a welfare list of class \code{welfare} returned by \code{\link[=welfare.change]{welfare.change()}}}

\item{label}{optional label to display}

\item{...}{further arguments passed to or from other methods}
}
\value{
The welfare list (invisibly).
}
\description{
\code{print.welfare()} provides a formatted welfare decomposition, showing
changes in consumer surplus, producer surplus, government revenue, and more.
}
