% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_welfare.R
\name{print_welfare}
\alias{print_welfare}
\title{Print a formatted welfare decomposition}
\usage{
print_welfare(x, label = "Welfare Change", ...)
}
\arguments{
\item{x}{a welfare list of class \code{welfare} returned by \code{\link[=welfare_change]{welfare_change()}}}

\item{label}{optional label to display}

\item{...}{further arguments passed to or from other methods}
}
\value{
The welfare list (invisibly).
}
\description{
\code{print_welfare()} provides a formatted welfare decomposition, showing
changes in consumer surplus, producer surplus, government revenue, and more.
}
