% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tariff_model.R
\name{print.welfare}
\alias{print.welfare}
\title{Print a formatted welfare decomposition}
\usage{
\method{print}{welfare}(wf, label = "Welfare Change")
}
\arguments{
\item{wf}{A welfare list returned by \code{welfare.change()}.}

\item{label}{Optional label to display.}
}
\description{
Print a formatted welfare decomposition
}
